fix: address pr feedback
